requirements.txt
aiogram>=3.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-socketio>=5.9.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
chess>=1.10.0
pydantic>=2.5.0
